requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
openai-whisper==20231117
python-dotenv==1.0.0
websockets==12.0
numpy==1.26.0
pydantic==2.4.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-socketio==5.9.0
eventlet==0.33.3
pydub==0.25.1
ffmpeg-python==0.2.0
